requirements.txt
asgiref==3.7.2
Django==4.1.12
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
pymongo==3.12.3
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
djangorestframework-simplejwt==5.3.0